+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Unreleased
 ### Added
 - Support for variable declarations.
+- Support for struct types.
 
 ### Fixed
 - Fix blocking assignments, making them actually block.